requirements.txt
absl-py==0.8.0
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.25.0
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.2
oauthlib==3.1.0
opt-einsum==3.1.0
protobuf==3.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.2
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
tensorboard==2.0.1
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
Unidecode==1.1.1
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
